budget_tracker_backend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
psycopg2==2.9.5
pytz==2022.7
sqlparse==0.4.3
